# docs/architecture/api/openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "本机管理系统当前 HTTP 契约",
    "version": "2026-08-08-working-tree",
    "description": "从当前Flask路由提取。x-legacy-exception表示请求/响应尚未统一；字段继续以源码与测试为准。"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:5000",
      "description": "默认本机Flask"
    }
  ],
...
  "paths": {
...
    "/api/browser-v2/elements": {
      "get": {
        "tags": [
          "Execution V2"
        ],
        "operationId": "execution_v2_list_elements_get",
        "summary": "GET /api/browser-v2/elements",
        "x-source": "execution_v2/blueprint.py:102",
        "x-legacy-exception": false,
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/GenericSuccess"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/Validation"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        }
      },
      "post": {
        "tags": [
          "Execution V2"
        ],
        "operationId": "execution_v2_create_element_post",
        "summary": "POST /api/browser-v2/elements",
        "x-source": "execution_v2/blueprint.py:106",
        "x-legacy-exception": false,
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/GenericSuccess"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/Validation"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        },
        "requestBody": {
          "required": false,
          "description": "具体字段以当前源码和测试为准；新改动必须补全严格Schema。",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      }
    },
    "/api/browser-v2/elements/{element_id}": {
      "get": {
        "tags": [
          "Execution V2"
        ],
        "operationId": "execution_v2_get_element_get",
        "summary": "GET /api/browser-v2/elements/{element_id}",
        "x-source": "execution_v2/blueprint.py:114",
        "x-legacy-exception": false,
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/GenericSuccess"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/Validation"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        },
        "parameters": [
          {
            "name": "element_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "put": {
        "tags": [
          "Execution V2"
        ],
        "operationId": "execution_v2_update_element_put",
        "summary": "PUT /api/browser-v2/elements/{element_id}",
        "x-source": "execution_v2/blueprint.py:118",
        "x-legacy-exception": false,
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/GenericSuccess"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/Validation"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        },
        "parameters": [
          {
            "name": "element_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": false,
          "description": "具体字段以当前源码和测试为准；新改动必须补全严格Schema。",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Execution V2"
        ],
        "operationId": "execution_v2_delete_element_delete",
        "summary": "DELETE /api/browser-v2/elements/{element_id}",
        "x-source": "execution_v2/blueprint.py:129",
        "x-legacy-exception": false,
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/GenericSuccess"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/Validation"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        },
        "parameters": [
          {
            "name": "element_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": false,
          "description": "具体字段以当前源码和测试为准；新改动必须补全严格Schema。",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      }
    },
    "/api/browser-v2/elements/{element_id}/validate": {
      "post": {
        "tags": [
          "Execution V2"
        ],
        "operationId": "execution_v2_validate_element_post",
        "summary": "POST /api/browser-v2/elements/{element_id}/validate",
        "x-source": "execution_v2/blueprint.py:136",
        "x-legacy-exception": false,
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/GenericSuccess"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/Validation"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        },
        "parameters": [
          {
            "name": "element_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": false,
          "description": "具体字段以当前源码和测试为准；新改动必须补全严格Schema。",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      }
    },
    "/api/browser-v2/profiles": {
      "get": {
        "tags": [
          "Execution V2"
        ],
        "operationId": "execution_v2_list_profiles_get",
        "summary": "GET /api/browser-v2/profiles",
        "x-source": "execution_v2/blueprint.py:142",
        "x-legacy-exception": false,
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/GenericSuccess"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/Validation"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        }
      }
    },
    "/api/browser-v2/content-libraries": {
      "get": {
        "tags": [
          "Execution V2"
        ],
        "operationId": "execution_v2_list_content_libraries_get",
        "summary": "GET /api/browser-v2/content-libraries",
        "x-source": "execution_v2/blueprint.py:146",
        "x-legacy-exception": false,
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/GenericSuccess"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/Validation"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        }
      }
    },
    "/api/browser-v2/picker/start": {
      "post": {
        "tags": [
          "Execution V2"
        ],
        "operationId": "execution_v2_start_picker_post",
        "summary": "POST /api/browser-v2/picker/start",
        "x-source": "execution_v2/blueprint.py:150",
        "x-legacy-exception": false,
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/GenericSuccess"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/Validation"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        },
        "requestBody": {
          "required": false,
          "description": "具体字段以当前源码和测试为准；新改动必须补全严格Schema。",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      }
    },
    "/api/browser-v2/picker/{session_id}": {
      "get": {
        "tags": [
          "Execution V2"
        ],
        "operationId": "execution_v2_get_picker_get",
        "summary": "GET /api/browser-v2/picker/{session_id}",
        "x-source": "execution_v2/blueprint.py:159",
        "x-legacy-exception": false,
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/GenericSuccess"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/Validation"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        },
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/browser-v2/picker/{session_id}/save": {
      "post": {
        "tags": [
          "Execution V2"
        ],
        "operationId": "execution_v2_save_picker_selection_post",
        "summary": "POST /api/browser-v2/picker/{session_id}/save",
        "x-source": "execution_v2/blueprint.py:163",
        "x-legacy-exception": false,
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/GenericSuccess"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/Validation"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        },
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": false,
          "description": "具体字段以当前源码和测试为准；新改动必须补全严格Schema。",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      }
    },
    "/api/browser-v2/picker/{session_id}/finish": {
      "post": {
        "tags": [
          "Execution V2"
        ],
        "operationId": "execution_v2_finish_picker_post",
        "summary": "POST /api/browser-v2/picker/{session_id}/finish",
        "x-source": "execution_v2/blueprint.py:183",
        "x-legacy-exception": false,
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/GenericSuccess"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/Validation"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        },
        "parameters": [
          {
            "name": "session_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": false,
          "description": "具体字段以当前源码和测试为准；新改动必须补全严格Schema。",
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": true
              }
            }
          }
        }
      }
    },
...
    "/api/browser-v2/wheel-calibration": {
      "get": {
        "tags": [
          "Execution V2"
        ],
        "operationId": "execution_v2_get_wheel_calibration_get",
        "summary": "GET /api/browser-v2/wheel-calibration",
        "x-source": "execution_v2/blueprint.py:193",
        "x-legacy-exception": false,
        "security": [
          {
            "cookieAuth": []
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/GenericSuccess"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "409": {
            "$ref": "#/components/responses/Conflict"
          },
          "422": {
            "$ref": "#/components/responses/Validation"
          },
          "500": {
            "$ref": "#/components/responses/Internal"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "session",
        "description": "认证模式Flask Session；local-direct模式由loopback/Host guard保护。"
      }
    },
...
    "responses": {
      "GenericSuccess": {
        "description": "当前成功响应；Legacy路由形状可能不统一。",
        "content": {
          "application/json": {
            "schema": {}
          }
        }
      },
      "BadRequest": {
        "description": "请求无效",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "Unauthorized": {
        "description": "未认证",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "Forbidden": {
        "description": "权限或本机边界拒绝",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
...
      "Conflict": {
        "description": "revision或状态冲突",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "Validation": {
        "description": "业务校验失败",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
      "Internal": {
        "description": "内部错误，消息固定且脱敏",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      }
    }
  }
}